requirements.txt
discogs_api
spotipy
tidalapi
python3-discogs-client